fixing error in slurm_check_account and slurm_check_user
run_command previously assumed that you would pass it a 255 length array for its optional output.
This was not actually the case in slurm_check_account and slurm_check_user, which lead to a segfault
in the event of more than 32 characters being attempting to be read into the buffer by run_command.
This adds a size parameter to the run_command function.