Adding in sorting to filter_group
Due to changes in scope we now have to account for having more
than just -lab groups. But we want to make sure that if a user
is a member of a -lab group that their default account is a -lab
group. We now psuedo sort the filtered group list to make sure
that the first element is a -lab group if the user is in one.

This also fixes a small memory leak due to not freeing regexes.